Move processor tests to AndroidEventProcessorTest
